requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic-settings==2.5.2
httpx==0.27.2
google-generativeai==0.8.5
psycopg[binary]==3.2.1
pytest==8.3.3
